#32 Refactored LandInfo job and extracted utils method to a separate class
